refactor: move ext. repo. inter. into domain layer
