+++ src/main/java/io/pakland/mdas/githubstats/domain/repository/OrganizationExternalRepository.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.infrastructure.rest.repository.ports;
+package io.pakland.mdas.githubstats.domain.repository;
 
 import io.pakland.mdas.githubstats.application.dto.OrganizationDTO;
 import io.pakland.mdas.githubstats.application.exceptions.HttpException;
