+++ src/main/java/io/pakland/mdas/githubstats/infrastructure/rest/repository/adapters/TeamGitHubRepository.java
@@ -4,7 +4,7 @@ import io.pakland.mdas.githubstats.application.dto.TeamDTO;
 import io.pakland.mdas.githubstats.application.dto.UserDTO;
 import io.pakland.mdas.githubstats.application.exceptions.HttpException;
 import io.pakland.mdas.githubstats.infrastructure.rest.repository.WebClientConfiguration;
-import io.pakland.mdas.githubstats.infrastructure.rest.repository.ports.TeamExternalRepository;
+import io.pakland.mdas.githubstats.domain.repository.TeamExternalRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.reactive.function.client.WebClientResponseException;
